+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+## [2.10.7] - 2021-06-29
+
 ### Changed
 
 - Core: only consider lines from the same supplier as the behavior component
+
+## [2.10.6] - 2021-06-25
 
 ### Fixed